backend/requirements.txt
Django==2.2.19
djangorestframework==3.12.4
python-dotenv==0.20.0
djoser==2.1.0
drf-yasg==1.21.7
django-filter==2.4.0
Pillow==9.2.0
drf-extra-fields==3.4.1
psycopg2-binary==2.9
gunicorn==20.0.4
